move normer
