+++ resources/assets/js/helpers/toNormedFreeResponseQuestionOptions.ts
